Complete change of principle, focus on multiple locks and summary
Use of hash tables for a current state of RCU locks, preconditions and loc that may have a problem. Lock names are kept for accessing the values in hash tables, because lock names are used as keys.
Operations for hash maps and sets are mostly implemented. Problem loc needs more operation and there is a need to make operations cooperate. This version is not in a operational state,
the rework needs to finish implementing all the operations as well as use them. Testing is required.